Moving `ReportingPeriodCalculator` to the `core` module Adding `DataRetentionCheckTask` to schedule removal of data Removing configs that are no longer used
