+++ api/src/main/resources/application.yml
@@ -20,7 +20,6 @@ api:
   data-processor:
     csv: com.lantanagroup.link.thsa.GenericCSVProcessor
     xlsx: com.lantanagroup.link.thsa.GenericXLSXProcessor
-  events:
   cors:
     allowed-origins: '*'
     allowed-headers: '*'
@@ -33,8 +32,6 @@ api:
       - OPTIONS
   document-reference-system: urn:ietf:rfc:3986
   socket-timeout: 30000
-  user:
-    timezone: UTC
   report-defs:
     max-retry: 5
     retry-wait: 5000
@@ -48,12 +45,12 @@ api:
 # every 2 minutes: */30 * * * * *"
 # 21st day of month at 1:11p PT: "0 11 20 3 * *"
 schedule:
-#  query-patient-list-cron: '0 */3 * * * *'  # every three hours
-#  generate-report:
+#  query-patient-list-cron: '0 0 */3 * * *'  # every three hours
+#  generate-and-submit-report:
 #    - measure-ids:
 #        - NHSNGlycemicControlHypoglycemicInitialPopulation
 #      reporting-period-method: LastMonth
-#      cron: '0 * * * * *'
+#      cron: '0 0 3 * * *'
 #      regenerate-if-exists: false
 
 # For querying one or more List resources from the FHIR server to use as a patient list for one or more census'
@@ -62,15 +59,12 @@ schedule:
 #  lists:
 #    - list-id: 1234
 #      census-identifier: NHSNGlycemicControlHypoglycemicInitialPopulation
+
 thsa:
   dataMeasureReportId: inventorydata
 query:
   require-https: true
   query-class: com.lantanagroup.link.query.uscore.Query
-  #  proxy-address: ''
-  allowed-remote:
-    - 0:0:0:0:0:0:0:1
-    - 127.0.0.1
 uscore:
   fhir-server-base:
   patient-resource-types:
