--- core/src/main/java/com/lantanagroup/link/config/api/ApiQueryConfigModes.java
@@ -1,13 +1,0 @@
-package com.lantanagroup.link.config.api;
-
-public enum ApiQueryConfigModes {
-  /**
-   * The API runs the queries.
-   */
-  Local,
-
-  /**
-   * The queries are run by the QueryAPI deployment.
-   */
-  Remote
-}
