--- core/src/main/java/com/lantanagroup/link/config/api/UserConfig.java
@@ -1,12 +1,0 @@
-package com.lantanagroup.link.config.api;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class UserConfig  {
-  /**
-   * <strong>api.user.timezone</strong><br>The timezone that should be used to store and search for data.
-   */
-  private String timezone;
-}
